created another component just to store player stuff
