--- client/src/components/BlackjackGame/style.css
@@ -1,40 +1,0 @@
-#House,
-#Player1 {
-    text-align: center;
-    margin: 10px;
-}
-
-#pointsHouse,
-#pointsPlayer1,
-#scoreHouse,
-#scorePlayer1,
-#House,
-#Player1 {
-    color:white;
-}
-
-#cardOneHouse,
-#cardTwoHouse,
-#cardOnePlayer1,
-#cardTwoPlayer1,
-.hitCardHouse,
-.hitCardPlayer1 {
-    width:100px;
-    height:150px;
-}
-
-#pointDisplay,
-#pointDisplayHousePlayer,
-#scoresDisplay,
-#scoresDisplayHousePlayer {
-    margin: 5px;
-    padding: 5px;
-
-}
-
-#roundOver,
-#winnerDisplay {
-    margin: 10px;
-    padding: 10px;
-
-}
