+++ client/src/components/BlackjackTable/index.js
@@ -4,28 +4,35 @@ import { Container } from 'reactstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import BlackjackDirections from "../BlackjackDirections";
 import BlackjackGame from "../BlackjackGame";
-// import BlackjackScoreCard from "../BlackjackScoreCard";
+import BlackjackScoreCard from "../BlackjackScoreCard";
 
 function BlackjackTable() {
-    const [directions, setDirections] = useState(true);
+    const [displayDirections, setDisplayDirections] = useState(true);
     const [startGame, setStartGame] = useState(false);
+    const [displayScoreCard, setDisplayScoreCard] = useState(false);
 
     function onStart() {
         console.log("clicking start btn");
         setStartGame(true);
-        setDirections(false);
+        setDisplayDirections(false);
+    }
+
+    function endRound() {
+        console.log("end of round triggered");
+        setDisplayScoreCard(true);
     }
 
     return (
         <Container id="blackjackTable">
             <h2>Single Player Blackjack</h2>
             <br />
-            {directions
+            {displayDirections
             && <BlackjackDirections 
             onStart={onStart}
             />}
-            {/* <BlackjackScoreCard /> */}
-            {/* have this card hidden until the end of game is called */}
+            {displayScoreCard
+            && <BlackjackScoreCard /> 
+            }
             {startGame 
             && <BlackjackGame /> 
             }
